course: sort students in handout same as for assignment (same as selected sorter)
